fix(delivery,console,runner): what the code-review found before the PR
Six findings from `/code-review` against `upstream/main`, four of them
documented-standard breaches this repo treats as bugs.

**Three docs describe a layout this branch superseded.**

`delivery/README.md` said comments ride the milestone list "and only it",
and that the field is "SERVED, not yet rendered: no console surface reads
it today". Both stopped being true: `Get` carries them for the one issue
the list can never reach, and the Validation tile renders the newest one.
`services/AGENTS.md` — "a README describing a superseded layout is a bug".

`lexicon.md` said the run-wide line speaks "only while they have nothing
to say", listed its whole two-sentence vocabulary, and required a trailing
`…` on anything in flight. The line now carries agent prose throughout a
run and is led by a pulse. `lexicon.md:17` makes amending it in the same
PR the rule, and says explicitly that it governs agent prose too — so the
entry now states the SHAPE asked of that prose (one line, present tense,
about the run and not a criterion) and leaves the words to the agent,
which is the whole point of the feature.

`liveLine.ts`'s header recorded the rejected alternative — a `phase` event
the runner emits — and this change looks like that alternative winning. It
is not, and the header now says why: a phase enum claims the SAME fact as
the rows at the same granularity from a second source, where the status
line is prose at a different granularity the skill tells the agent to keep
disjoint from them. The header also has to say it is the fallback now.

`sourcecontrol/README.md` gains the new port beside its sibling, with the
distinction that matters: both are GraphQL, for different reasons. The
milestone read because REST cannot answer N issues in one call; the single
issue because REST answers it from the wrong end.

**The two comment reads had drifted into two copies of one shape.**

`issue_comments.go` repeated `milestone_comments.go`'s node struct and
projection verbatim. A comment's shape is stated three times — selection
set, decode struct, projection — and held twice that is six places that
must agree. `comments.go` now holds one of each, shared: `commentSelection`
is embedded in both queries, so the fields requested and the fields
decoded cannot drift; `commentNode` and `toIssueComments` are the decode
and the mapping. `isMachineComment` and `maxCommentsPerIssue` move there
too, since both readers already used them.

**`Get` took its comment round trip in sequence.**

The list read fires its comments alongside the issue query and records the
measurement for doing so (614ms → 1345ms sequential, 614ms → 775ms
together, on a 5s poll). `Get` is now polled on exactly that cadence while
a run is in flight and was taking a second host round trip strictly after
the first. It now overlaps, the same way and for the same reason.

Deliberately NOT made opt-in like the list's `?comments=`. That flag exists
because a milestone-wide read has no bounded set of issues to ask about;
one issue is bounded by construction, and with the fetch overlapped the
caller that renders no comments now pays no latency for it.

**`chown -R /ms-playwright` was a whole second copy of the browsers.**

7db0050a put it in the RUN that creates the user, which is after the one
that downloads them — and a recursive chown rewrites every file it touches
into a fresh layer. The Dockerfile already records that exact cost eighty
lines down, on the app COPY: "a trailing `chown -R /app` … would rewrite
every file as a fresh layer and roughly double the image size". Two
chromium revisions is not a rounding error, on an image `runners/AGENTS.md`
already flags as GC-eviction-sensitive.

The user is created BEFORE the Playwright block now, so the install layer
writes the ownership itself and there is no second copy.

**The skill named a step that asks for no comment.**

`## The status line` said steps 1, 2, 4 and 10 ask for comments. Step 4
asks the agent to READ a comment the platform posted on a different issue —
the roles gate ticket — so listing it as one of this line's ends pointed
at the wrong issue entirely. The two ends are step 1's opening comment and
step 10's closing summary.

Left alone, with reasons:

- `FULL_WIDTH_ALERT_MESSAGE` exported from `LiveNote.tsx` (flagged as
  Divergent Change). It exists because LiveNote is a panel rather than
  text, and living beside it means deleting LiveNote reveals it is unused.
- `const v = validationScenario()` in the mock handler (flagged as a
  Mysterious Name). Four existing call sites in that file already read that
  way; the repo's local idiom overrides the baseline.
- The pulse assertion living in `ValidationPage.test.tsx` rather than a
  `LiveNote.test.tsx`. Proving it renders in situ is the stronger test.

Verified: full Go suite, 1766 console tests, 449 runner tests, tsc, eslint,
license-check, deadcode-check (Go), deadcode-ts-check, `go vet`, gofmt.
